+++ PdLib/src/x_acoustics.c
@@ -13,12 +13,12 @@ t_float mtof(t_float f)
 {
     if (f <= -1500) return(0);
     else if (f > 1499) return(mtof(1499));
-    else return (8.17579891564 * exp(.0577622650 * f));
+    else return (8.17579891564 * expf(.0577622650 * f));
 }
 
 t_float ftom(t_float f)
 {
-    return (f > 0 ? 17.3123405046 * log(.12231220585 * f) : -1500);
+    return (f > 0 ? 17.3123405046 * logf(.12231220585 * f) : -1500);
 }
 
 t_float powtodb(t_float f)
@@ -26,7 +26,7 @@ t_float powtodb(t_float f)
     if (f <= 0) return (0);
     else
     {
-        t_float val = 100 + 10./LOGTEN * log(f);
+        t_float val = 100 + 10./LOGTEN * logf(f);
         return (val < 0 ? 0 : val);
     }
 }
@@ -36,7 +36,7 @@ t_float rmstodb(t_float f)
     if (f <= 0) return (0);
     else
     {
-        t_float val = 100 + 20./LOGTEN * log(f);
+        t_float val = 100 + 20./LOGTEN * logf(f);
         return (val < 0 ? 0 : val);
     }
 }
@@ -49,7 +49,7 @@ t_float dbtopow(t_float f)
     {
         if (f > 870)
             f = 870;
-        return (exp((LOGTEN * 0.1) * (f-100.)));
+        return (expf((LOGTEN * 0.1) * (f-100.)));
     }
 }
 
@@ -62,7 +62,7 @@ t_float dbtorms(t_float f)
         if (f > 485)
             f = 485;
     }
-    return (exp((LOGTEN * 0.05) * (f-100.)));
+    return (expf((LOGTEN * 0.05) * (f-100.)));
 }
 
 /* ------------- corresponding objects ----------------------- */
